pyproject.toml
[tool.poetry]
name = "nearestneighbors"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
  { include = "ui", from = "src" },
  { include = "llm", from = "src" },
  { include = "nn", from = "src" },
  { include = "utils", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.12"
polars = "^1.16.0"
pandas = "^2.2.3"
scikit-learn = "^1.5.2"
streamlit = "^1.40.2"
jupyterlab = "^4.3.1"
openpyxl = "^3.1.5"
llama-index = "^0.12.2"
llama-index-llms-groq = "^0.3.0"
groq = "^0.13.0"
python-dotenv = "^1.0.1"
matplotlib = "^3.9.3"
plotly = "^5.24.1"
duckdb = "^1.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

